requirements.txt
django==3.2
psycopg2-binary==2.9.1
graphene-django==3.1.5
graphql-core==3.2.3
redis==3.5.3
requests==2.26.0
djangorestframework==3.12.4
celery==5.2.0
